トラック生成機能を別クラスに分離した
ローカル関数をメンバー関数に

ローカル関数を外に出し中

同上

同上

引数整理中

引数を調整

同上

同上

途中
